moving function, class and instance into grammar dir
