+++ src/grammar/mod.rs
@@ -1,5 +1,8 @@
 pub mod callable;
+pub mod class;
 pub mod expr;
+pub mod function;
+pub mod instance;
 pub mod native_function;
 pub mod object;
 pub mod stmt;
